fix bugs in simple gl application
The texture containing the 3D scene needed coordinate transforms to be
properly displayed. Also, the properties of the camera class should be
known to the QML engine: even if real is a typedef on double and qreal
is another typedef on double, real is unknown to QML while qreal is
known. Finaly, all the fields in camera are now initialized to correct
values.