working on analyzing OpenCL devices.
Some information about the OpenCL devices available on the computer are logged
at the initialization of the tools library. There is still work needed to
improve the quality of the information displayed as well as to handle correctly
devices/platforms: on my computer, a platform is not correctly displayed and I
believe this is the ATI graphic card.